pyproject.toml
[tool.poetry]
name = "agent-eval-framework"
version = "0.1.0"
description = "A framework for evaluating agents"
authors = ["laah laah@google.com"]
license = "PRIVATE"
readme = "README.md"
packages = [{include = "agent_eval_framework", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
google-cloud-aiplatform = { extras = ["evaluation", "rapid_evaluation"], version = "^1.113.0" }
google-cloud-storage = "^2.18.0"
google-cloud-trace = "1.14.0"
google-auth = "~=2.32.0"
ipython = "^8.26.0"

# OpenTelemetry
opentelemetry-api = "^1.31.0"
opentelemetry-sdk = "^1.31.0"
opentelemetry-exporter-gcp-trace = "~=1.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-mock = "^3.15.0"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src", "."
]
